fix: unify dtype to BF16 across Pi0VLA and PaliGemmaBackbone
pi0_core.py self.half() overrode backbone BF16, causing dtype mismatch.
Now both use bfloat16(). Backbone forward reads dtype from parameters
instead of hardcoding, so it stays consistent.

+++ models/pi0_core.py
@@ -61,8 +61,8 @@ class Pi0VLA(nn.Module):
         )
 
 
-        # ── 3. 전체 FP16 (Jetson AGX) ─────────────────────────────────
-        self.half()
+        # BF16 (FP16은 NaN 유발)
+        self.bfloat16()
 
     # ─────────────────────────────────────────────────────────────────
     def forward_backbone(self, images: torch.Tensor, instructions) -> torch.Tensor:
